\name{randomMatrix}
\alias{randomMatrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Random Transition Matrix Generation
}
\description{
Generates a random 22x22 transition matrix for use with markovMelodies
}
\usage{
randomMatrix()
}
%- maybe also 'usage' for other objects documented here.
\arguments{

}
\details{
Function requires no arguments, and is affected by set.seed()
}
\value{
Square 22x22 float matrix.
}
\references{
https://github.com/chnnxyz/markov.melodies
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{randomWalk}},\code{\link{secondThirdFifthHD}}
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
